refactor(snapshot): correct the name of `ProcessSnapshot` class
I had originally named it `RequireSnapshot` and I didn't quite finish the rename to `ProcessSnapshot` before committing.
